requirements.txt
Flask>=2.2
Flask-SQLAlchemy>=3.0
PyJWT>=2.8
Werkzeug>=2.2
openai>=1.0.0
